refactor(web): extract useKeyboardShortcuts from App.tsx
Mechanical move of isEditableTarget, the handleGlobalKeydown
useEffectEvent, and the keydown listener effect into
hooks/useKeyboardShortcuts.ts with dependencies injected via a single
options object. No shortcut semantics change.

Issue: CS-18 (step 2/3)
